+++ app/static/lib/control-menu.js
@@ -1,5 +1,5 @@
 import * as icon from './../css/icons.js';
-import { fontList, platform } from './main.js';
+import { fontList,platform } from './defaults.js';
 import { svgNameSpace } from './dom-utils.js';
 import { createPageRangeSelector } from './page-range-selector.js';
 
@@ -455,6 +455,30 @@ export function createFacsimileControlBar(parentElement) {
   fullPageDiv.appendChild(fullPageCheckbox);
 
   facsCtrlBar.appendChild(fullPageDiv);
+
+  // show zone rectangles
+  let showZonesDiv = document.createElement('div');
+  showZonesDiv.id = 'facsimile-show-zones';
+  showZonesDiv.classList.add('controls');
+
+  let showZonesLabel = document.createElement('label');
+  showZonesLabel.innerText = 'Show zone boxes:';
+  showZonesLabel.id = 'facsimile-show-zones-label';
+  showZonesLabel.classList.add('label');
+  showZonesDiv.appendChild(showZonesLabel);
+  showZonesLabel.title = 'Show zone boxes of facsimile';
+
+  let showZonesCheckbox = document.createElement('input');
+  showZonesCheckbox.id = 'facsimile-show-zones-checkbox';
+  showZonesCheckbox.setAttribute('type', 'checkbox');
+  showZonesCheckbox.classList.add('checkbox');
+  showZonesCheckbox.title = 'Show zone boxes of facsimile';
+  showZonesLabel.setAttribute('for', showZonesCheckbox.id);
+  showZonesCheckbox.checked = false;
+  showZonesCheckbox.disabled = false;
+  showZonesDiv.appendChild(showZonesCheckbox);
+
+  facsCtrlBar.appendChild(showZonesDiv);
 
   // edit zones
   let editZonesDiv = document.createElement('div');