+++ app/static/lib/viewer.js
@@ -9,15 +9,9 @@ import { drawFacsimile, highlightZone, zoomFacsimile } from './facsimile.js';
 import {
   cm,
   cmd,
-  commonSchemas,
-  defaultVerovioVersion,
-  fontList,
-  isSafari,
   meiFileName,
   rngLoader,
-  platform,
   storage,
-  supportedVerovioVersions,
   tkVersion,
   validate,
   validator,
@@ -27,6 +21,18 @@ import {
 import { selectMarkup } from './markup.js';
 import { mp, startMidiTimeout } from './midi-player.js';
 import { getNotationProportion, setNotationProportion, setOrientation } from './resizer.js';
+import {
+  commonSchemas,
+  codeMirrorSettingsOptions,
+  defaultNotationProportion,
+  defaultSpeedMode,
+  defaultViewerTimeoutDelay,
+  fontList,
+  isSafari,
+  meiFriendSettingsOptions,
+  platform,
+  supportedVerovioVersions,
+} from './defaults.js';
 
 export default class Viewer {
   constructor(vrvWorker, spdWorker) {
@@ -44,7 +50,7 @@ export default class Viewer {
     this.notationNightMode = false;
     this.allowCursorActivity = true; // whether or not notation gets re-rendered after text changes
     this.allowNotationInteraction = true; // allow mouse drag-select and click on notation
-    this.speedMode = true; // speed mode (just feeds on page to Verovio to reduce drawing time)
+    this.speedMode = defaultSpeedMode; // speed mode (just feeds on page to Verovio to reduce drawing time)
     this.parser = new DOMParser();
     this.xmlDoc;
     this.xmlDocOutdated = true; // to limit recalculation of DOM or pageLists
@@ -59,7 +65,7 @@ export default class Viewer {
     this.meiHeadRange = [];
     this.vrvOptions; // all verovio options
     this.vrvTimeout; // time out task for updating verovio settings
-    this.timeoutDelay = 300; // ms, window in which concurrent clicks are treated as one update
+    this.timeoutDelay = defaultViewerTimeoutDelay; // ms, window in which concurrent clicks are treated as one update
     this.verovioIcon = document.getElementById('verovio-icon');
     this.breaksSelect = /** @type HTMLSelectElement */ (document.getElementById('breaks-select'));
     this.respId = '';
@@ -68,7 +74,7 @@ export default class Viewer {
     this.cmd2KeyPressed = false;
     this.controlMenuState = {};
     this.settingsReplaceFriendContainer = false; // whether or not the settings panel is over the mei-friend window (false) or replaces it (true)
-    this.notationProportion = 0.5; // remember proportion during pdf mode
+    this.notationProportion = defaultNotationProportion; // remember proportion during pdf mode
     this.expansionId = ''; // id string of currently selected expansion element
   } // constructor() // constructor()
 
@@ -445,7 +451,7 @@ export default class Viewer {
     }
     utils.setCursorToId(cm, id);
     // console.info('setCrsrToPgBeg(): lastNoteId: ' + this.lastNoteId + ', new id: ' + id);
-    this.selectedElements.push(id);
+    if (!this.selectedElements.includes(id)) this.selectedElements.push(id);
     this.lastNoteId = id;
     return id;
   } // setCursorToPageBeginning()
@@ -520,7 +526,7 @@ export default class Viewer {
       // console.log('cursorActivity forceFlip: ' + forceFlip + ' to: ' + id);
       this.selectedElements = [];
       if (id) {
-        this.selectedElements.push(id);
+        if (!this.selectedElements.includes(id)) this.selectedElements.push(id);
         let fl = document.getElementById('flip-checkbox');
         if (
           !document.querySelector('g#' + utils.escapeXmlId(id)) && // when not on current page
@@ -1003,438 +1009,7 @@ export default class Viewer {
   } // applySettingsFilter()
 
   addMeiFriendOptionsToSettingsPanel(restoreFromLocalStorage = true) {
-    let meiFriendSettingsOptions = {
-      titleGeneral: {
-        title: 'General',
-        description: 'General mei-friend settings',
-        type: 'header',
-        default: true,
-      },
-      selectToolkitVersion: {
-        title: 'Verovio version',
-        description:
-          'Select Verovio toolkit version ' +
-          '(* Switching to older versions before 3.11.0 ' +
-          'might require a refresh due to memory issues.)',
-        type: 'select',
-        default: defaultVerovioVersion,
-        values: Object.keys(supportedVerovioVersions),
-        valuesDescriptions: Object.keys(supportedVerovioVersions).map((key) => {
-          let desc = supportedVerovioVersions[key].description;
-          if (supportedVerovioVersions[key].hasOwnProperty('releaseDate')) {
-            desc += ' (' + supportedVerovioVersions[key].releaseDate + ')';
-          }
-          return desc;
-        }),
-      },
-      toggleSpeedMode: {
-        title: 'Speed mode',
-        description:
-          'Toggle Verovio Speed Mode. ' +
-          'In speed mode, only the current page ' +
-          'is sent to Verovio to reduce rendering ' +
-          'time with large files',
-        type: 'bool',
-        default: true,
-      },
-      selectIdStyle: {
-        title: 'Style of generated xml:ids',
-        description:
-          'Style of newly generated xml:ids (existing xml:ids are not changed)' +
-          'e.g., Verovio original: "note-0000001318117900", ' +
-          'Verovio base 36: "nophl5o", ' +
-          'mei-friend style: "note-ophl5o"',
-        type: 'select',
-        values: ['Original', 'Base36', 'mei-friend'],
-        valuesDescriptions: ['note-0000001018877033', 'n34z4wz2', 'note-34z4wz2'],
-        default: 'Base36',
-      },
-      addApplicationNote: {
-        title: 'Insert application statement',
-        description:
-          'Insert an application statement to the encoding ' +
-          'description in the MEI header, identifying ' +
-          'application name, version, date of first ' +
-          'and last edit',
-        type: 'bool',
-        default: true,
-      },
-      dragSelection: {
-        title: 'Drag select',
-        description: 'Select elements in notation with mouse drag',
-        type: 'header',
-        default: true,
-      },
-      dragSelectNotes: {
-        title: 'Select notes',
-        description: 'Select notes',
-        type: 'bool',
-        default: true,
-      },
-      dragSelectRests: {
-        title: 'Select rests',
-        description: 'Select rests and repeats (rest, mRest, beatRpt, halfmRpt, mRpt)',
-        type: 'bool',
-        default: true,
-      },
-      dragSelectControlElements: {
-        title: 'Select placement elements ',
-        description:
-          'Select placement elements (i.e., with a @placement attribute: ' + att.attPlacement.join(', ') + ')',
-        type: 'bool',
-        default: false,
-      },
-      dragSelectSlurs: {
-        title: 'Select slurs ',
-        description: 'Select slurs (i.e., elements with @curvature attribute: ' + att.attCurvature.join(', ') + ')',
-        type: 'bool',
-        default: false,
-      },
-      dragSelectMeasures: {
-        title: 'Select measures ',
-        description: 'Select measures',
-        type: 'bool',
-        default: false,
-      },
-      controlMenuSettings: {
-        title: 'Notation control bar',
-        description: 'Define items to be shown in control menu above the notation',
-        type: 'header',
-        default: true,
-      },
-      // flip-checkbox, flip-btn
-      controlMenuFlipToPageControls: {
-        title: 'Show flip to page controls',
-        description: 'Show flip to page controls in notation control menu',
-        type: 'bool',
-        default: true,
-      },
-      controlMenuUpdateNotation: {
-        title: 'Show notation update controls',
-        description: 'Show notation update behavior controls in notation control menu',
-        type: 'bool',
-        default: true,
-      },
-      controlMenuFontSelector: {
-        title: 'Show notation font selector',
-        description: 'Show notation font (SMuFL) selector in notation control menu',
-        type: 'bool',
-        default: false,
-      },
-      controlMenuNavigateArrows: {
-        title: 'Show navigation arrows',
-        description: 'Show notation navigation arrows in notation control menu',
-        type: 'bool',
-        default: false,
-      },
-      controlMenuSpeedmodeCheckbox: {
-        title: 'Show speed mode checkbox',
-        description: 'Show speed mode checkbox in notation control menu',
-        type: 'bool',
-        default: true,
-      },
-      titleMidiPlayback: {
-        title: 'MIDI playback',
-        description: 'MIDI playback settings',
-        type: 'header',
-        default: true,
-      },
-      showMidiPlaybackContextualBubble: {
-        title: 'Show playback shortcut',
-        description:
-          'Causes a shortcut (bubble in bottom left corner; click to immediately start playback) to appear when the MIDI playback control bar is closed',
-        type: 'bool',
-        default: true,
-      },
-      showMidiPlaybackControlBar: {
-        title: 'Show MIDI playback control bar',
-        description: 'Show MIDI playback control bar',
-        type: 'bool',
-        default: false,
-      },
-      scrollFollowMidiPlayback: {
-        title: 'Scroll-follow MIDI playback',
-        description: 'Scroll notation panel to follow MIDI playback on current page',
-        type: 'bool',
-        default: true,
-      },
-      pageFollowMidiPlayback: {
-        title: 'Page-follow MIDI playback',
-        description: 'Automatically flip pages to follow MIDI playback',
-        type: 'bool',
-        default: true,
-      },
-      titleTransposition: {
-        title: 'Transpose',
-        description: 'Transpose score information',
-        type: 'header',
-        default: true,
-      },
-      enableTransposition: {
-        title: 'Enable transposition',
-        description:
-          'Enable transposition settings, to be applied through the transpose button below. The transposition will be applied to the notation only, the encoding remains unchanged, unless you click the item "Rerender via Verovio" in the "Manipulate" dropdown menu.',
-        type: 'bool',
-        default: false,
-      },
-      transposeInterval: {
-        title: 'Transpose by interval',
-        description:
-          'Transpose encoding by chromatic interval by the most common intervals (Verovio supports the base-40 system)',
-        type: 'select',
-        labels: [
-          'Perfect Unison',
-          'Augmented Unison',
-          'Diminished Second',
-          'Minor Second',
-          'Major Second',
-          'Augmented Second',
-          'Diminished Third',
-          'Minor Third',
-          'Major Third',
-          'Augmented Third',
-          'Diminished Fourth',
-          'Perfect Fourth',
-          'Augmented Fourth',
-          'Diminished Fifth',
-          'Perfect Fifth',
-          'Augmented Fifth',
-          'Diminished Sixth',
-          'Minor Sixth',
-          'Major Sixth',
-          'Augmented Sixth',
-          'Diminished Seventh',
-          'Minor Seventh',
-          'Major Seventh',
-          'Augmented Seventh',
-          'Diminished Octave',
-          'Perfect Octave',
-        ],
-        values: [
-          'P1',
-          'A1',
-          'd2',
-          'm2',
-          'M2',
-          'A2',
-          'd3',
-          'm3',
-          'M3',
-          'A3',
-          'd4',
-          'P4',
-          'A4',
-          'd5',
-          'P5',
-          'A5',
-          'd6',
-          'm6',
-          'M6',
-          'A6',
-          'd7',
-          'm7',
-          'M7',
-          'A7',
-          'd8',
-          'P8',
-        ],
-        default: 'P1',
-        radioId: 'byInterval',
-        radioName: 'transposeMode',
-      },
-      transposeKey: {
-        title: 'Transpose to key',
-        description: 'Transpose to key',
-        type: 'select',
-        labels: [
-          'C# major / A# minor',
-          'F# major / D# minor',
-          'B major / G# minor',
-          'E major / C# minor',
-          'A major / F# minor',
-          'D major / B minor',
-          'G major / E minor',
-          'C major / E minor',
-          'F major / D minor',
-          'Bb major / G minor',
-          'Eb major / C minor',
-          'Ab major / F minor',
-          'Db major / Bb minor',
-          'Gb major / Eb minor',
-          'Cb major / Ab minor',
-        ],
-        values: ['cs', 'fs', 'b', 'e', 'a', 'd', 'g', 'c', 'f', 'bf', 'ef', 'af', 'df', 'gf', 'cf'],
-        default: 'c',
-        radioId: 'toKey',
-        radioName: 'transposeMode',
-        radioChecked: 'true',
-      },
-      transposeDirection: {
-        title: 'Pitch direction',
-        description: 'Pitch direction of transposition (up/down)',
-        type: 'select',
-        labels: ['Up', 'Down', 'Closest'],
-        values: ['+', '-', ''],
-        default: '+', // refers to values
-      },
-      transposeButton: {
-        title: 'Transpose',
-        description:
-          'Apply transposition with above settings to the notation, while the MEI encoding remains unchanged. To also transpose the MEI encoding with the current settings, use "Rerender via Verovio" in the "Manipulate" dropdown menu.',
-        type: 'button',
-      },
-      renumberMeasuresHeading: {
-        title: 'Renumber measures',
-        description: 'Settings for renumbering measures',
-        type: 'header',
-        default: true,
-      },
-      renumberMeasureContinueAcrossIncompleteMeasures: {
-        title: 'Continue across incomplete measures',
-        description: 'Continue measure numbers across incomplete measures (@metcon="false")',
-        type: 'bool',
-        default: false,
-      },
-      renumberMeasuresUseSuffixAtMeasures: {
-        title: 'Suffix at incomplete measures',
-        description: 'Use number suffix at incomplete measures (e.g., 23-cont)',
-        type: 'select',
-        values: ['none', '-cont'],
-        default: false,
-      },
-      renumberMeasuresContinueAcrossEndings: {
-        title: 'Continue across endings',
-        description: 'Continue measure numbers across endings',
-        type: 'bool',
-        default: false,
-      },
-      renumberMeasuresUseSuffixAtEndings: {
-        title: 'Suffix at endings',
-        description: 'Use number suffix at endings (e.g., 23-a)',
-        type: 'select',
-        values: ['none', 'ending@n', 'a/b/c', 'A/B/C', '-a/-b/-c', '-A/-B/-C'],
-        default: 'a/b/c',
-      },
-      // annotationPanelSeparator: {
-      //   title: 'options-line', // class name of hr element
-      //   type: 'line'
-      // },
-      highlightCurrentlySoundingNotes: {
-        title: 'Highlight currently-sounding notes',
-        description: 'Visually highlight currently-sounding notes in the notation panel during MIDI playback ',
-        type: 'bool',
-        default: true,
-      },
-      selectMidiExpansion: {
-        title: 'Expansion for MIDI playback',
-        description: 'Select expansion element to be used for MIDI playback',
-        type: 'select',
-        values: ['none'],
-        // valuesDescriptions: ['note-0000001018877033', 'n34z4wz2', 'note-34z4wz2'],
-        default: 'none',
-      },
-      titleAnnotations: {
-        title: 'Annotations',
-        description: 'Annotation settings',
-        type: 'header',
-        default: true,
-      },
-      showAnnotations: {
-        title: 'Show annotations',
-        description: 'Show annotations in notation',
-        type: 'bool',
-        default: true,
-      },
-      showAnnotationPanel: {
-        title: 'Show annotation panel',
-        description: 'Show annotation panel',
-        type: 'bool',
-        default: false,
-      },
-      annotationDisplayLimit: {
-        title: 'Maximum number of annotations',
-        description: 'Maximum number of annotations to display ' + '(large numbers may slow mei-friend)',
-        type: 'int',
-        min: 0,
-        step: 100,
-        default: 100,
-      },
-      titleFacsimilePanel: {
-        title: 'Facsimile panel',
-        description: 'Show the facsimile imiages of the source edition, if available',
-        type: 'header',
-        open: false,
-        default: false,
-      },
-      showFacsimilePanel: {
-        title: 'Show facsimile panel',
-        description: 'Show the score images of the source edition provided in the facsimile element',
-        type: 'bool',
-        default: false,
-      },
-      selectFacsimilePanelOrientation: {
-        title: 'Facsimile panel position',
-        description: 'Select facsimile panel position relative to notation',
-        type: 'select',
-        values: ['left', 'right', 'top', 'bottom'],
-        default: 'bottom',
-      },
-      facsimileZoomInput: {
-        title: 'Facsimile image zoom (%)',
-        description: 'Zoom level of facsimile image (in percent)',
-        type: 'int',
-        min: 10,
-        max: 300,
-        step: 5,
-        default: 100,
-      },
-      showFacsimileFullPage: {
-        title: 'Show full page',
-        description: 'Show facsimile image on full page',
-        type: 'bool',
-        default: false,
-      },
-      editFacsimileZones: {
-        title: 'Edit facsimile zones',
-        description: 'Edit facsimile zones (will link bounding boxes to facsimile zones)',
-        type: 'bool',
-        default: false,
-      },
-      // sourcefacsimilePanelSeparator: {
-      //   title: 'options-line', // class name of hr element
-      //   type: 'line'
-      // },
-      titleSupplied: {
-        title: 'Handle editorial content',
-        description: 'Control handling of <supplied> elements',
-        type: 'header',
-        open: false,
-        default: false,
-      },
-      showSupplied: {
-        title: 'Show <supplied> elements',
-        description: 'Highlight all elements contained by a <supplied> element',
-        type: 'bool',
-        default: true,
-      },
-      suppliedColor: {
-        title: 'Select <supplied> highlight color',
-        description: 'Select <supplied> highlight color',
-        type: 'color',
-        default: '#e69500',
-      },
-      respSelect: {
-        title: 'Select <supplied> responsibility',
-        description: 'Select responsibility id',
-        type: 'select',
-        default: 'none',
-        values: [],
-      },
-      // dragLineSeparator: {
-      //   title: 'options-line', // class name of hr element
-      //   type: 'line'
-      // },
-    };
+    // NOTE: object meiFriendSettingsOptions from defaults.js
     let mfs = document.getElementById('meiFriendSettings');
     let addListeners = false; // add event listeners only the first time
     let rt = document.querySelector(':root');
@@ -1517,6 +1092,9 @@ export default class Viewer {
           break;
         case 'editFacsimileZones':
           document.getElementById('facsimile-edit-zones-checkbox').checked = value;
+          break;
+        case 'showFacsimileZones':
+          document.getElementById('facsimile-show-zones-checkbox').checked = value;
           break;
         case 'showMidiPlaybackControlBar':
           // do nothing, as it is always the default display: none
@@ -1665,6 +1243,14 @@ export default class Viewer {
           case 'transposeInterval':
           case 'transposeDirection':
             break;
+          case 'showFacsimileZones':
+            document.getElementById('facsimile-show-zones-checkbox').checked = value;
+            if (!value) {
+              document.getElementById('editFacsimileZones').checked = false;
+              document.getElementById('facsimile-edit-zones-checkbox').checked = false;
+            }
+            drawFacsimile();
+            break;
           case 'selectMidiExpansion':
             this.updateSelectMidiExpansion();
             if (document.getElementById('showMidiPlaybackControlBar').checked) {
@@ -1673,6 +1259,10 @@ export default class Viewer {
             break;
           case 'editFacsimileZones':
             document.getElementById('facsimile-edit-zones-checkbox').checked = value;
+            if (value) {
+              document.getElementById('showFacsimileZones').checked = true;
+              document.getElementById('facsimile-show-zones-checkbox').checked = true;
+            }
             drawFacsimile();
             break;
           case 'showFacsimilePanel':
@@ -1793,143 +1383,7 @@ export default class Viewer {
   } // addMeiFriendOptionsToSettingsPanel()
 
   addCmOptionsToSettingsPanel(mfDefaults, restoreFromLocalStorage = true) {
-    let codeMirrorSettingsOptions = {
-      // key as in CodeMirror
-      titleAppearance: {
-        title: 'Editor appearance',
-        description: 'Controls the appearance of the editor',
-        type: 'header',
-        open: true,
-        default: true,
-      },
-      zoomFont: {
-        title: 'Font size (%)',
-        description: 'Change font size of editor (in percent)',
-        type: 'int',
-        default: 100,
-        min: 45,
-        max: 300,
-        step: 5,
-      },
-      theme: {
-        title: 'Theme',
-        description: 'Select the theme of the editor',
-        type: 'select',
-        default: 'default',
-        values: [
-          'default',
-          'abbott',
-          'base16-dark',
-          'base16-light',
-          'cobalt',
-          'darcula',
-          'dracula',
-          'eclipse',
-          'elegant',
-          'monokai',
-          'idea',
-          'juejin',
-          'mdn-like',
-          'neo',
-          'paraiso-dark',
-          'paraiso-light',
-          'pastel-on-dark',
-          'solarized dark',
-          'solarized light',
-          'xq-dark',
-          'xq-light',
-          'yeti',
-          'yonce',
-          'zenburn',
-        ],
-      },
-      matchTheme: {
-        title: 'Notation matches theme',
-        description: 'Match notation to editor color theme',
-        type: 'bool',
-        default: false,
-      },
-      tabSize: {
-        title: 'Indentation size',
-        description: 'Number of space characters for each indentation level',
-        type: 'int',
-        min: 1,
-        max: 12,
-        step: 1,
-        default: 3,
-      },
-      lineWrapping: {
-        title: 'Line wrapping',
-        description: 'Whether or not lines are wrapped at end of panel',
-        type: 'bool',
-        default: false,
-      },
-      lineNumbers: {
-        title: 'Line numbers',
-        description: 'Show line numbers',
-        type: 'bool',
-        default: true,
-      },
-      firstLineNumber: {
-        title: 'First line number',
-        description: 'Set first line number',
-        type: 'int',
-        min: 0,
-        max: 1,
-        step: 1,
-        default: 1,
-      },
-      foldGutter: {
-        title: 'Code folding',
-        description: 'Enable code folding through fold gutters',
-        type: 'bool',
-        default: true,
-      },
-      titleEditorOptions: {
-        title: 'Editor behavior',
-        description: 'Controls the behavior of the editor',
-        type: 'header',
-        open: true,
-        default: true,
-      },
-      autoValidate: {
-        title: 'Auto validation',
-        description: 'Validate encoding against schema automatically after each edit',
-        type: 'bool',
-        default: true,
-      },
-      autoCloseBrackets: {
-        title: 'Auto close brackets',
-        description: 'Automatically close brackets at input',
-        type: 'bool',
-        default: true,
-      },
-      autoCloseTags: {
-        title: 'Auto close tags',
-        description: 'Automatically close tags at input',
-        type: 'bool',
-        default: true,
-      },
-      matchTags: {
-        title: 'Match tags',
-        description: 'Highlights matched tags around editor cursor',
-        type: 'bool',
-        default: true,
-      },
-      showTrailingSpace: {
-        title: 'Highlight trailing spaces',
-        description: 'Highlights unnecessary trailing spaces at end of lines',
-        type: 'bool',
-        default: true,
-      },
-      keyMap: {
-        title: 'Key map',
-        description: 'Select key map',
-        type: 'select',
-        default: 'default',
-        values: ['default', 'vim', 'emacs'],
-      },
-    };
+    // NOTE: codeMirrorSettingsOptions in defaults.js
     let storage = window.localStorage;
     let cmsp = document.getElementById('editorSettings');
     let addListeners = false; // add event listeners only the first time
@@ -2447,7 +1901,7 @@ export default class Viewer {
     // this.allowCursorActivityToTextposition(txtEdr); TODO
     if (id) {
       this.selectedElements = [];
-      this.selectedElements.push(id);
+      if (!this.selectedElements.includes(id)) this.selectedElements.push(id);
       this.lastNoteId = id;
       if (document.getElementById('showMidiPlaybackControlBar').checked) {
         startMidiTimeout();
@@ -2637,13 +2091,13 @@ export default class Viewer {
     if (disappearAfter > 0) {
       this.alertCloser = setTimeout(() => (alertOverlay.style.display = 'none'), disappearAfter);
     }
-  }
+  } // showAlert()
 
   // Update alert message of #alertOverlay
   updateAlert(newMsg) {
     let alertOverlay = document.getElementById('alertOverlay');
-    alertOverlay.querySelector('span').innerHTML += '<br />' + newMsg;
-  }
+    alertOverlay.querySelector('span#alertMessage').innerHTML += '<br />' + newMsg;
+  } // updateAlert()
 
   // Hide all alert windows, such as alert overlay
   hideAlerts() {
@@ -2652,7 +2106,7 @@ export default class Viewer {
       if (this.alertCloser) clearTimeout(this.alertCloser);
       b.parentElement.style.display = 'none';
     }
-  }
+  } // hideAlerts()
 
   // Method to check from MEI whether the XML schema filename has changed
   async checkSchema(mei) {